+++ test/NestedSubstitutionSpec.hs
@@ -15,7 +15,7 @@ import API.Types (RootDb (..), Substitution (..), SubstitutionScope (..), Supply
 import qualified Data.Map.Strict as M
 import qualified Data.Text as T
 import qualified Data.Vector.Unboxed as U
-import Matrix (buildDemandVectorFromIndex)
+import Matrix (buildDemandVector)
 import Service (
     ActivityFilterCore (..),
     Edges (..),
@@ -54,7 +54,7 @@ spec = do
             db <- loadSampleDatabase "SAMPLE.min3"
             solver <- mkSolver db "root"
             let pid = 0
-                demandVec = buildDemandVectorFromIndex (dbActivityIndex db) pid
+                demandVec = buildDemandVector db pid
             baselineX <- solveWithSharedSolver solver =<< either (fail . show) pure demandVec
             let qualifiedPid = "other-db::aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa_bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"
                 sub =
@@ -76,7 +76,7 @@ spec = do
             db <- loadSampleDatabase "SAMPLE.min3"
             solver <- mkSolver db "root"
             let pid = 0
-                demandVec = buildDemandVectorFromIndex (dbActivityIndex db) pid
+                demandVec = buildDemandVector db pid
             baselineX <- solveWithSharedSolver solver =<< either (fail . show) pure demandVec
             let noDeps _ = pure Nothing
             res <- applySubstitutionsAt defaultUnitConfig noDeps db (ThisDb "root") (RootDb "root") solver [baselineX] []
@@ -93,7 +93,7 @@ spec = do
             db <- loadSampleDatabase "SAMPLE.min3"
             solver <- mkSolver db "dep"
             let pid = 0
-                demandVec = buildDemandVectorFromIndex (dbActivityIndex db) pid
+                demandVec = buildDemandVector db pid
             baselineX <- solveWithSharedSolver solver =<< either (fail . show) pure demandVec
             let qualifiedToRoot = "root::aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa_bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"
                 sub =
@@ -117,7 +117,7 @@ spec = do
             db <- loadSampleDatabase "SAMPLE.min3"
             solver <- mkSolver db "dep"
             let pid = 0
-                demandVec = buildDemandVectorFromIndex (dbActivityIndex db) pid
+                demandVec = buildDemandVector db pid
             baselineX <- solveWithSharedSolver solver =<< either (fail . show) pure demandVec
             let bareRootPid = "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa_bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"
                 sub =
@@ -274,7 +274,7 @@ spec = do
             depSolver <- mkSolverFromDb dep "dep"
             let lookup_ = mkDepLookupFromMap (M.singleton "dep" (dep, depSolver))
                 pid = 0
-                demandVec = buildDemandVectorFromIndex (dbActivityIndex root) pid
+                demandVec = buildDemandVector root pid
             scaling <- solveWithSharedSolver rootSolver =<< either (fail . show) pure demandVec
             let scf =
                     SupplyChainFilter